requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
attrs
certifi
charset-normalizer
colorama
Django
django-cors-headers
frozenlist
gunicorn
idna
multidict
openai
packaging
psycopg2
requests
sqlparse
tqdm
typing_extensions
tzdata
urllib3
yarl
google-generativeai